refactor: 폼 관련 상수들을 목적별로 분리하여 구조 개선
- validation.ts에서 폼 옵션 관련 상수들을 분리
- form-options.ts 파일 신규 생성 (GENDER_OPTIONS, DEFAULT_AGREEMENTS)
- 각 상수 파일의 책임을 명확히 분리:
  - validation.ts: 유효성 검사 관련 상수만 유지
  - form-options.ts: 폼에서 사용하는 옵션들
  - beauty-profile.ts: 뷰티프로필 관련 상수들
  - form-styles.ts: 폼 스타일 관련 상수들
- profile-edit/page.tsx에서 새로운 import 경로로 수정
- 코드의 가독성과 유지보수성 향상
